# src/docs/swagger.yaml
basePath: /
definitions:
  types.AudioRequest:
    properties:
      audio:
        type: string
    type: object
  types.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  types.GetInfoResponse:
    properties:
      task_id:
        type: string
    type: object
  types.GetResultResponse:
    properties:
      result:
        type: string
    type: object
  types.GetStatusResponse:
    properties:
      status:
        type: string
    type: object
  types.PaginationResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  types.TaskInfo:
    properties:
      created:
        type: string
      status:
        type: string
      task_id:
        type: string
      username:
        type: string
    type: object
  types.TaskListResponse:
    properties:
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
      tasks:
        items:
          $ref: '#/definitions/types.TaskInfo'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Speech to Text API
  title: Speech to Text API
  version: "1.0"
paths:
  /audio:
    post:
      consumes:
      - application/json
      description: Sends audio file for speech to text conversion
      parameters:
      - description: Audio data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.AudioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task ID created
          schema:
            $ref: '#/definitions/types.GetInfoResponse'
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Upload audio for processing
      tags:
      - audio
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates user and creates session
      parameters:
      - description: Authentication data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful authentication
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User authentication
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful registration
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User registration
      tags:
      - auth
  /result:
    get:
      consumes:
      - application/json
      description: Returns audio to text conversion result
      parameters:
      - description: Task ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.GetInfoResponse'
      produces:
      - application/json
      responses:
        "200":
          description: Processing result
          schema:
            $ref: '#/definitions/types.GetResultResponse'
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get processing result
      tags:
      - tasks
  /status:
    get:
      consumes:
      - application/json
      description: Returns current audio processing status
      parameters:
      - description: Task ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.GetInfoResponse'
      produces:
      - application/json
      responses:
        "200":
          description: Task status
          schema:
            $ref: '#/definitions/types.GetStatusResponse'
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get task status
      tags:
      - tasks
  /tasks:
    get:
      consumes:
      - application/json
      description: Returns user tasks list with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks list
          schema:
            $ref: '#/definitions/types.TaskListResponse'
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get tasks list with pagination
      tags:
      - tasks
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"